Move the Script node classes to their own folder.
