+++ lib/sass/script/nodes/interpolation.rb
