+++ lib/sass/script/nodes/string_interpolation.rb
