+++ lib/sass/script/nodes/variable.rb
